+++ ui/mods/com.pa.quitch.gwaioverhaul/gw_play/section_of_foreign_intelligence/section_of_foreign_intelligence.js
@@ -245,8 +245,7 @@ function gwoIntelligence() {
             gwoCards.anyPlayerHasCard(inventory, "gwaio_enable_suddendeath")
           ) {
             gameModifiers.push(loc("!LOC:Sudden Death"));
-          }
-          if (
+          } else if (
             ai.eradicationMode ||
             gwoCards.anyPlayerHasCard(inventory, "gwaio_enable_eradication")
           ) {